renamed new_model to model
